requirements.txt
# Core dependencies
langchain>=0.0.267
langchain-openai
langchain-community>=0.0.10
langchain-experimental
openai>=1.3.0
python-dotenv>=1.0.0

# Document processing
unstructured>=0.10.0
pdfminer.six>=20221105
pymupdf>=1.22.5
unstructured_inference>=0.6.3
unstructured_pytesseract>=0.3.12
pillow_heif>=0.10.0
pi_heif>=0.13.0

# Output formatting
pandas>=2.0.0
xlsxwriter>=3.0.0

# Azure specific
azure-identity>=1.13.0
azure-storage-blob>=12.17.0

